% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSDIF.R
\name{tosdif}
\alias{tosdif}
\title{TOSDIF}
\usage{
tosdif(
  fit1,
  fit2,
  groups = NULL,
  alpha = NULL,
  rmseaa = NULL,
  rmseab = NULL,
  robust = TRUE
)
}
\arguments{
\item{fit1}{First fit-object to be compared}

\item{fit2}{Second fit-object to be compared}

\item{groups}{Number of groups in the model. Defaults to 1 if not specified}

\item{alpha}{Significance level for the difference in fit. Defaults to 0.05 if not specified (recommended)}

\item{rmseaa}{Upper limit used to define a small difference in fit. Defaults to 0.06 if not specified (recommended)}

\item{rmseab}{Lower limit used to define a small difference in fit. Defaults to 0.05 if not specified (recommended)}

\item{robust}{Used to determine if robust version of Chi-Square should be used. Only available, if fit-objects were estimated using "estimator = MLR" or similar}
}
\value{
p-value for test of small differences in fit and interpretation
}
\description{
Test of small differences in fit for two fit-objects.
}
